pyproject.toml
[project]
name = "paul-graham-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.1.0",
    "google-generativeai>=0.8.4",
    "ipykernel>=6.29.5",
    "ipython>=8.32.0",
    "llama-index>=0.12.16",
    "llama-index-embeddings-huggingface>=0.5.1",
    "llama-index-llms-gemini>=0.4.10",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "streamlit>=1.42.0",
]
